# swagger.yaml
openapi: 3.0.3
info:
  title: Frames API
  description: API for managing picture frames, walls, and 3D models
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://your-production-url.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
  schemas:
...
    Wall:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        description:
          type: string
        image_path:
          type: string
          nullable: true
        thumbnail_path:
          type: string
          nullable: true
        background_color:
          type: string
          nullable: true
        width_cm:
          type: number
          format: float
          nullable: true
        height_cm:
          type: number
          format: float
          nullable: true
        scene_config:
          type: object
        frame_placements:
          type: array
          items:
            $ref: '#/components/schemas/FramePlacement'
        created_at:
          type: string
          format: date-time
    FramePlacement:
      type: object
      properties:
        frame_id:
          type: integer
        picture_id:
          type: integer
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            z:
              type: number
        rotation:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
            z:
              type: number
        scale:
          type: number
          format: float
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - BearerAuth: []
paths:
...
  /walls:
    get:
      tags:
        - Walls
      summary: Get all walls
      description: Retrieve all walls for the authenticated user
      responses:
        '200':
          description: List of walls
          content:
            application/json:
              schema:
                type: object
                properties:
                  walls:
                    type: array
                    items:
                      $ref: '#/components/schemas/Wall'

    post:
      tags:
        - Walls
      summary: Create a new wall
      description: Create a wall from uploaded image or solid color
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Living Room Wall
                description:
                  type: string
                  example: Main wall in living room
                width_cm:
                  type: number
                  format: float
                  example: 400.0
                height_cm:
                  type: number
                  format: float
                  example: 250.0
                background_color:
                  type: string
                  example: "#FFFFFF"
                image:
                  type: string
                  format: binary
                  description: Wall image file (png, jpg, jpeg, gif, webp)
      responses:
        '201':
          description: Wall created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  wall:
                    $ref: '#/components/schemas/Wall'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /walls/{wall_id}:
    get:
      tags:
        - Walls
      summary: Get a specific wall
      description: Retrieve a wall by ID with its assigned frames
      parameters:
        - name: wall_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Wall details
          content:
            application/json:
              schema:
                type: object
                properties:
                  wall:
                    $ref: '#/components/schemas/Wall'
        '404':
          description: Wall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Walls
      summary: Update a wall
      description: Update wall details or frame placements
      parameters:
        - name: wall_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                width_cm:
                  type: number
                  format: float
                height_cm:
                  type: number
                  format: float
                background_color:
                  type: string
                scene_config:
                  type: object
                frame_placements:
                  type: array
                  items:
                    $ref: '#/components/schemas/FramePlacement'
      responses:
        '200':
          description: Wall updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  wall:
                    $ref: '#/components/schemas/Wall'
        '404':
          description: Wall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Walls
      summary: Delete a wall
      description: Delete a wall and its associated files
      parameters:
        - name: wall_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Wall deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Wall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /walls/{wall_id}/placements:
    post:
      tags:
        - Walls
      summary: Add frame placement
      description: Add a frame placement to a wall
      parameters:
        - name: wall_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FramePlacement'
      responses:
        '200':
          description: Frame placement added
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  wall:
                    $ref: '#/components/schemas/Wall'
        '404':
          description: Wall not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...